Created a base class called MovingProvider for later usage for the settings section
