+++ lib/providers/nav_provider.dart
@@ -1,10 +1,9 @@
-import 'package:flutter/material.dart';
 import 'package:hive/hive.dart';
 
-class NavProvider extends ChangeNotifier {
-  bool isShown = false; // Default value
+import 'moving_provider.dart';
+
+class NavProvider extends MovingProvider {
   int selectedIndex = 0;
-  late Box settingsBox; // Hive box for storing settings
 
   NavProvider() {
     _loadSettings(); // Load settings from Hive on initialization
@@ -12,23 +11,18 @@ class NavProvider extends ChangeNotifier {
 
   // Load settings from Hive
   void _loadSettings() async {
-    settingsBox = await Hive.openBox('settings');
-    isShown = settingsBox.get('isShown', defaultValue: false);
-    selectedIndex = settingsBox.get('selectedIndex', defaultValue: 0);
+    box = await Hive.openBox('navigation_settings');
+    isShown = box.get('isShown', defaultValue: false);
+    selectedIndex = box.get('selectedIndex', defaultValue: 0);
 
     notifyListeners(); // Notify listeners to rebuild UI with the loaded theme
   }
 
+  // Update index to selected value
   void setIndex(value) {
     selectedIndex = value;
-    settingsBox.put('selectedIndex', selectedIndex);
-    notifyListeners();
-  }
+    box.put('selectedIndex', selectedIndex);
 
-  // Function to toggle between showing the navigational rails and not
-  void toggleNavBar() {
-    isShown = !isShown;
-    settingsBox.put('isShown', isShown); // Save the new value to Hive
-    notifyListeners(); // Notify listeners to rebuild UI
+    notifyListeners();
   }
 }
